# projects/omni-api-realtime-platform/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Omni-Channel Real-Time Enterprise API
  version: 1.0.0
  description: |
    Unified production-grade REST API adhering to Enterprise Standards:
    - Kebab-case resource endpoints with max 2-level hierarchy.
    - Idempotency-Key header support preventing double mutations.
    - RFC 7807 Problem Details for all client and server errors.
    - Cursor & Keyset pagination with IETF Link header conventions.
  contact:
    name: Enterprise Architecture Team
    email: api-core@enterprise.internal
servers:
  - url: http://localhost:4000/api/v1
    description: Local Development Environment
  - url: https://api.enterprise.internal/v1
    description: Production Gateway
paths:
  /orders:
    get:
      summary: List Orders with Keyset Pagination and Filters
      description: Retrieve orders sorted chronologically with cursor-based pagination.
      operationId: listOrders
      tags:
        - Orders
      parameters:
        - name: limit
          in: query
          description: Maximum number of records to return (1-100).
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          description: Opaque cursor string for the next page.
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filter orders by lifecycle status.
          required: false
          schema:
            type: string
            enum: [PENDING, PROCESSING, COMPLETED, CANCELLED]
      responses:
        '200':
          description: Paginated list of orders retrieved successfully.
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request allowance per minute.
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining quota in current window.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '500':
          $ref: '#/components/responses/InternalErrorProblem'

    post:
      summary: Create Order with Idempotency Key
      description: Places a new customer order. Safe to retry with the same `Idempotency-Key` header.
      operationId: createOrder
      tags:
        - Orders
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Unique client-generated UUIDv4/UUIDv7 to guarantee at-most-once execution.
          schema:
            type: string
            format: uuid
            example: "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully.
          headers:
            Location:
              schema:
                type: string
                example: "/api/v1/orders/ord_987654321"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/InternalErrorProblem'
  /orders/{orderId}:
    get:
      summary: Get Order by ID
      operationId: getOrderById
      tags:
        - Orders
      parameters:
        - name: orderId
          in: path
          required: true
          description: Unique order identifier.
          schema:
            type: string
            example: "ord_101"
      responses:
        '200':
          description: Order details retrieved.
          headers:
            ETag:
              schema:
                type: string
                example: 'W/"ord_101_v3"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFoundProblem'

    patch:
      summary: Partial Update Order (JSON Merge Patch RFC 7396)
      operationId: patchOrder
      tags:
        - Orders
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
        - name: If-Match
          in: header
          required: true
          description: ETag for optimistic concurrency control.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [PENDING, PROCESSING, COMPLETED, CANCELLED]
      responses:
        '200':
          description: Order updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '412':
          description: Precondition Failed - ETag mismatch.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Zero-Trust Cryptographic JWT containing sub, iss, aud, and scope claims.
  schemas:
    Order:
      type: object
      required:
        - id
        - customerId
        - items
        - totalAmount
        - currency
        - status
        - createdAt
      properties:
        id:
          type: string
          example: "ord_101"
        customerId:
          type: string
          example: "cust_55"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: double
          example: 249.99
        currency:
          type: string
          example: "USD"
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, CANCELLED]
          example: "PROCESSING"
        createdAt:
          type: string
          format: date-time
          example: "2026-09-14T00:00:00Z"
        version:
          type: integer
          example: 1
    OrderItem:
      type: object
      required:
        - productId
        - quantity
        - unitPrice
      properties:
        productId:
          type: string
          example: "prod_laptop_01"
        quantity:
          type: integer
          minimum: 1
          example: 1
        unitPrice:
          type: number
          format: double
          example: 249.99
    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
      properties:
        customerId:
          type: string
          example: "cust_55"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderPage:
      type: object
      required:
        - items
        - pageInfo
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pageInfo:
          type: object
          properties:
            hasNextPage:
              type: boolean
            endCursor:
              type: string
            totalCount:
              type: integer
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details Standard Specification
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: "https://api.enterprise.internal/errors/bad-request"
        title:
          type: string
          example: "Invalid Request Payload"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "Field 'items' must contain at least 1 item."
        instance:
          type: string
          example: "/api/v1/orders"
        invalidParams:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
  responses:
    BadRequestProblem:
      description: Bad Request (RFC 7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFoundProblem:
      description: Resource Not Found (RFC 7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ConflictProblem:
      description: State Conflict / Duplicate Execution (RFC 7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalErrorProblem:
      description: Internal Server Error (RFC 7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
security:
  - BearerAuth: []